web/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
qrcode==7.4.2
Pillow==10.1.0
requests==2.31.0
websockets==11.0.3
redis==5.0.1
celery==5.3.4
python-socketio==5.10.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
